Solve tasks for pypi exercise
